+++ src/js/Bounds.js
@@ -10,7 +10,8 @@
  * @author   Ikaros Kappler
  * @date     2020-05-11
  * @modified 2020-10-30 Added the static computeFromVertices function.
- * @version  1.1.0
+ * @modified 2020-11-19 Set min, max, width and height to private.
+ * @version  1.1.1
  *
  * @file Bounds
  * @fileoverview A simple bounds class implementing IBounds.