+++ src/ts/Bounds.ts
@@ -9,7 +9,8 @@
  * @author   Ikaros Kappler
  * @date     2020-05-11
  * @modified 2020-10-30 Added the static computeFromVertices function.
- * @version  1.1.0
+ * @modified 2020-11-19 Set min, max, width and height to private.
+ * @version  1.1.1
  *
  * @file Bounds
  * @fileoverview A simple bounds class implementing IBounds.
@@ -27,7 +28,7 @@ export class Bounds implements IBounds, XYDimension {
      * @instance
      * @public
      */
-    min:XYCoords;
+    readonly min:XYCoords;
 
     /** 
      * @member {XYCoords} 
@@ -35,7 +36,7 @@ export class Bounds implements IBounds, XYDimension {
      * @instance
      * @public
      */
-    max:XYCoords;
+    readonly max:XYCoords;
 
     /** 
      * @member {number} 
@@ -43,7 +44,7 @@ export class Bounds implements IBounds, XYDimension {
      * @instance
      * @public
      */
-    width:number;
+    readonly width:number;
 
     /** 
      * @member {number} 
@@ -51,7 +52,7 @@ export class Bounds implements IBounds, XYDimension {
      * @instance
      * @public
      */
-    height:number;
+    readonly height:number;