+++ changelog.md
@@ -1,13 +1,15 @@
 
 
 # Changelog
+* 2020-11-19
+  * Set min, max, width and height to private.
 * 2020-11-17
   * Added the `VertexListeners.click(VertListener)` handler.
   * Added pure click handling (no dragEnd and !wasMoved jiggliny any more) to the PlotBoilerplate.
 * 2020-11-11
   * Generalized `move(Vertex)` to `move(XYCoords)` in the Polygon class.
   * Generalized the `VertTuple.add` and `VertTuple.sub` param from `Vertex` to `XYCoords`.
-  * Girih-demo: ported the GirihTile class from vanilla JS to TypeScript.	
+  * Girih-demo: ported the GirihTile class from vanilla JS to TypeScript.
 * 2020-11-10
   * Added the `Polygon.getBounds()` function.
 * 2020-11-06